+++ packages/react/src/components/FormField/src/formField.types.ts
@@ -31,7 +31,13 @@ export interface FormFieldProperties {
    * Appearance of the field and help text elements.
    * @default none
    */
-  appearance?: "danger" | "warning" | "success" | "none";
+  appearance?:
+    | "danger"
+    | "warning"
+    | "success"
+    | "none"
+    | "ai-generative"
+    | "ai-generated";
   /**
    * Control to conditionally show the help text and icon.
    * @default false